Architecture:
I have parsed all of the useful CSS out of the in-line styles for
the most part, and that now lives in public_style.css

I have cleaned the About page (about.html) dutifully and fixed up some
known issues with that web page.  Fixed console errors on that page,
organized dependencies, and added and reformatted on the javascript code.

ToDo:
Get the other pages up to speed with About
Move the javascript to a separate file
Organize JS dependencies for common architecture instead of redundancy
More CSS and fixing things up
Make common methodology for printing header and footer.